+++ HISTORY.rst
@@ -2,8 +2,22 @@
 History
 =======
 
-0.1.0 (2020-03-24)
-------------------
+**0.1.2 (2020-04-23)**
+
+Additional functionality.
+
+* Added functions for temporal aggregation (daily to monthly, daily to yearly)
+* Statistics of simulated discharge (mean, max, min etc) can be computed
+
+**0.1.1 (2020-04-20)**
+
+Minor updates.
+
+* Moved towards object based programming
+* Reading from 'normal' csv-files supported now
+
+**0.1.0 (2020-03-24)**
+
 First stable release with minimal functionality.
 
 * Extracting timeseries from netCDF-files for certain lon/lat